requirements.txt
google-generativeai
chromadb
langchain
langchain-community
langchain-text-splitters
tqdm
python-dotenv
streamlit
docling
PyMuPDF